more robust unit test error reporting

+++ test/app-under-test.js
@@ -12,11 +12,10 @@ module.exports = {
         var opts = {
             cwd : __dirname + "/.."
         };
-        
-        child = c.fork("index", args, opts);
-            
+
         var deferred = Q.defer();
         
+        child = c.fork("index", args, opts);
         child.on('message', function(data) {
             deferred.resolve();
         });
@@ -24,15 +23,23 @@ module.exports = {
         return deferred.promise;
     },
     
+    // idempotent
     stop : function() {
         var deferred = Q.defer();
-        
-        child.on('exit', function() {
-            deferred.resolve();
-        });
 
-        child.kill('SIGINT');
-       // child.kill('SIGTERM');
+        if (child) {
+            child.on('exit', function() {
+                deferred.resolve();
+            });
+
+            child.kill('SIGINT');
+            child = undefined;
+        } else {
+            setTimeout(function() {
+                deferred.resolve();
+            }, 0);
+        }
+
         return deferred.promise;
     }
 }
